fixed warning after upgrading to Xcode 5.1 the AFNetworking warning should be fixed by upgrading to the afnetworking version that addresses it

+++ Echowaves/EWWave.m
@@ -130,7 +130,7 @@
     manager.responseSerializer = [AFJSONResponseSerializer serializer];
     manager.requestSerializer = [AFJSONRequestSerializer serializer];
     
-    NSDictionary *parameters = @{@"badge": [NSString stringWithFormat: @"%d", numberOfImages],
+    NSDictionary *parameters = @{@"badge": [NSString stringWithFormat: @"%ld", (long)numberOfImages],
                                  @"wave_name": waveName};
     
     [manager POST:[NSString stringWithFormat:@"%@/send-push-notify.json", EWHost] parameters:parameters success:^(AFHTTPRequestOperation *operation, id responseObject) {
